Move metadataflyout to consistent folder
