+++ bundles/catalogue/metadataflyout/handler/MetadataHandler.js
