+++ bundles/catalogue/metadataflyout/handler/util.js
