+++ bundles/catalogue/metadataflyout/index.js
