+++ bundles/catalogue/metadataflyout/instance.js
