+++ bundles/catalogue/metadataflyout/request/ShowMetadataRequestHandler.js
