+++ bundles/catalogue/metadataflyout/resources/locale/en.js
