+++ bundles/catalogue/metadataflyout/resources/locale/fr.js
