+++ bundles/catalogue/metadataflyout/resources/locale/is.js
