+++ bundles/catalogue/metadataflyout/resources/locale/ru.js
