+++ bundles/catalogue/metadataflyout/view/MetadataContainer.jsx
