+++ bundles/catalogue/metadataflyout/view/MetadataFlyout.jsx
