+++ bundles/catalogue/metadataflyout/view/tab/components/Label.jsx
