+++ bundles/catalogue/metadataflyout/view/tab/components/LabeledItem.jsx
