+++ bundles/catalogue/metadataflyout/view/tab/index.js
